% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKLR.R
\name{predict.MKLR}
\alias{predict.MKLR}
\title{Predict values by MKLR model}
\usage{
\method{predict}{MKLR}(object, newx, response = c("probability", "class")[1])
}
\arguments{
\item{object}{MKLR model.}

\item{newx}{new data without labels}

\item{response}{probability or class.}
}
\value{
The predicted probabilities or classes.
}
\description{
Prediction
}
\seealso{
\link{MKLR}
}
